FINAL CLEANUP & REFACTOR
