+++ pwa/src/components/formFields/input.tsx
@@ -1,8 +1,8 @@
 import * as React from "react";
-import { InputGroup } from "./formGroup";
-import { IFormGroupProps, IReactHookFormProps } from "./types";
+import { InputGroup } from "./formFieldGroup";
+import { IFormFieldProps, IReactHookFormProps } from "./types";
 
-export const InputText: React.FC<IFormGroupProps & IReactHookFormProps> = ({
+export const InputText: React.FC<IFormFieldProps & IReactHookFormProps> = ({
   name,
   label,
   errors,
